folder structure simplified
