--- DOTA_devkit/.gitignore
@@ -1,13 +1,0 @@
-build
-polyiou_wrap.cxx
-.idea/DOTAPreprocess.iml
-*.so
-*.jpg
-**/.idea/
-*pyc
-examplesplit
-__pycache__/
-.ipynb_checkpoints/
-Task1_merge/*.txt
-Task1/*.txt
-restoredexample/*
