--- DOTA_devkit/example/labelTxt/P0770.txt
@@ -1,24 +1,0 @@
-imagesource:GoogleEarth
-gsd:0.266578848023
-856 696 1469 1008 1341 1298 716 988 ground-track-field 0
-1148 658 1255 711 1217 787 1110 735 swimming-pool 0
-349 491 388 508 353 591 313 573 tennis-court 0
-346 362 362 369 345 405 328 398 swimming-pool 0
-248 577 330 613 314 652 231 614 tennis-court 0
-543 66 560 76 540 115 521 104 swimming-pool 0
-283 491 306 490 305 534 282 533 swimming-pool 0
-244 643 292 664 278 694 231 672 swimming-pool 0
-210 686 254 704 247 720 204 702 swimming-pool 0
-223 1010 240 1018 226 1050 209 1043 swimming-pool 0
-14 1091 56 1108 45 1131 5 1112 swimming-pool 0
-125 1240 142 1247 130 1282 113 1275 swimming-pool 0
-1002 339 1040 360 999 440 962 420 tennis-court 0
-941 310 978 329 939 409 900 390 tennis-court 0
-879 279 918 299 878 379 839 359 tennis-court 0
-819 249 857 269 816 349 779 329 tennis-court 0
-757 219 795 238 756 319 718 299 tennis-court 0
-697 189 735 208 695 290 657 269 tennis-court 0
-638 351 781 422 716 567 565 498 baseball-diamond 0
-394 825 466 690 600 762 528 905 baseball-diamond 0
-653 795 747 845 698 936 602 890 tennis-court 0
-496 148 535 166 498 247 459 230 tennis-court 0
