--- DOTA_devkit/example/labelTxt/P1088.txt
@@ -1,36 +1,0 @@
-imagesource:GoogleEarth
-gsd:0.1377814038
-226 738 225 756 191 756 191 739 small-vehicle 0
-1033 545 948 547 949 490 1032 490 plane 0
-407 532 323 535 323 479 410 478 plane 0
-506 539 428 537 435 476 513 483 plane 0
-615 545 528 541 533 486 615 488 plane 0
-711 538 630 540 630 484 711 482 plane 0
-816 545 730 544 733 479 819 483 plane 0
-918 547 837 545 844 487 923 490 plane 0
-1242 25 1238 107 1178 107 1184 22 plane 0
-248 666 248 687 191 685 192 669 small-vehicle 0
-1121 286 1040 282 1045 223 1127 227 plane 0
-1016 278 942 276 944 225 1018 229 plane 0
-919 281 840 280 836 220 922 221 plane 0
-809 279 741 277 745 222 814 226 plane 0
-617 270 541 271 541 220 616 221 plane 0
-410 263 358 267 353 224 408 221 plane 0
-384 914 252 910 261 797 393 802 plane 0
-418 809 495 811 495 897 415 893 plane 0
-540 783 670 787 669 908 536 909 plane 0
-584 1165 689 1171 686 1282 576 1281 plane 0
-756 1173 882 1179 879 1301 755 1297 plane 0
-462 1664 630 1666 623 1831 467 1832 plane 0
-988 1609 1162 1608 1161 1808 986 1805 plane 0
-230 574 230 592 192 591 193 575 small-vehicle 0
-230 595 229 615 193 614 193 595 small-vehicle 0
-67 719 84 719 82 755 65 756 small-vehicle 0
-101 738 116 736 115 771 100 771 small-vehicle 0
-124 732 140 731 138 767 124 767 small-vehicle 0
-191 1103 208 1103 206 1138 191 1137 small-vehicle 0
-261 384 260 411 182 410 184 385 large-vehicle 0
-267 432 267 454 200 457 200 434 large-vehicle 0
-245 478 244 499 197 501 199 481 large-vehicle 0
-64 734 46 735 47 691 66 691 small-vehicle 0
-321 267 241 268 239 214 323 217 plane 0
