--- DOTA_devkit/example/labelTxt/P1234.txt
@@ -1,146 +1,0 @@
-imagesource:GF
-gsd:1
-2512 3597 2528 3602 2519 3642 2504 3637 bridge 0
-2598 2655 2615 2659 2605 2693 2584 2688 bridge 0
-2604 3931 2614 3931 2614 3942 2604 3942 storage-tank 1
-2593 3930 2603 3930 2603 3941 2593 3941 storage-tank 1
-2583 3927 2593 3927 2593 3939 2583 3939 storage-tank 1
-2571 3927 2580 3927 2580 3936 2571 3936 storage-tank 1
-2935 3089 2945 3089 2945 3098 2935 3098 storage-tank 1
-2946 3082 2953 3082 2953 3091 2946 3091 storage-tank 1
-2838 3206 2848 3206 2848 3216 2838 3216 storage-tank 1
-2847 3197 2860 3197 2860 3209 2847 3209 storage-tank 1
-2858 3191 2868 3191 2868 3202 2858 3202 storage-tank 1
-2868 3182 2879 3182 2879 3192 2868 3192 storage-tank 1
-2878 3173 2892 3173 2892 3185 2878 3185 storage-tank 1
-2894 3162 2906 3162 2906 3173 2894 3173 storage-tank 1
-2907 3152 2918 3152 2918 3163 2907 3163 storage-tank 1
-2911 3061 2924 3061 2924 3073 2911 3073 storage-tank 1
-2898 3045 2911 3045 2911 3058 2898 3058 storage-tank 1
-2615 3933 2626 3933 2626 3946 2615 3946 storage-tank 1
-2626 3937 2635 3937 2635 3947 2626 3947 storage-tank 1
-2637 3938 2648 3938 2648 3948 2637 3948 storage-tank 1
-2577 3957 2585 3957 2585 3967 2577 3967 storage-tank 1
-3150 3828 3166 3828 3166 3847 3150 3847 storage-tank 0
-2631 3968 2641 3968 2641 3978 2631 3978 storage-tank 1
-2621 3964 2630 3964 2630 3976 2621 3976 storage-tank 1
-2609 3963 2619 3963 2619 3973 2609 3973 storage-tank 1
-2599 3962 2608 3962 2608 3971 2599 3971 storage-tank 1
-2588 3960 2597 3960 2597 3968 2588 3968 storage-tank 1
-2565 3956 2574 3956 2574 3965 2565 3965 storage-tank 1
-2567 3940 2578 3940 2578 3950 2567 3950 storage-tank 1
-2635 3953 2644 3953 2644 3963 2635 3963 storage-tank 1
-2624 3951 2633 3951 2633 3962 2624 3962 storage-tank 1
-2612 3947 2622 3947 2622 3959 2612 3959 storage-tank 1
-2602 3946 2610 3946 2610 3956 2602 3956 storage-tank 1
-2591 3944 2600 3944 2600 3955 2591 3955 storage-tank 1
-2579 3943 2589 3943 2589 3952 2579 3952 storage-tank 1
-2751 3057 2762 3057 2762 3067 2751 3067 storage-tank 1
-2735 3053 2747 3053 2747 3064 2735 3064 storage-tank 1
-2740 3035 2752 3035 2752 3049 2740 3049 storage-tank 1
-1340 3378 1360 3378 1360 3400 1340 3400 storage-tank 0
-1791 3833 1803 3833 1803 3847 1791 3847 storage-tank 1
-2102 1823 2124 1840 1862 2233 1839 2217 bridge 0
-1400 1620 1414 1636 1341 1687 1328 1673 ship 0
-1174 1001 1179 996 1195 1011 1190 1017 ship 0
-1155 1103 1159 1098 1171 1112 1166 1115 ship 0
-1376 3395 1396 3395 1396 3417 1376 3417 storage-tank 0
-1303 3362 1326 3362 1326 3385 1303 3385 storage-tank 0
-1770 3824 1786 3824 1786 3837 1770 3837 storage-tank 0
-1311 3333 1323 3333 1323 3345 1311 3345 storage-tank 1
-1288 3313 1305 3313 1305 3330 1288 3330 storage-tank 0
-1270 3347 1289 3347 1289 3366 1270 3366 storage-tank 0
-1282 3434 1323 3434 1323 3475 1282 3475 storage-tank 0
-1232 3412 1272 3412 1272 3452 1232 3452 storage-tank 0
-956 672 961 668 969 682 964 685 ship 0
-1787 3850 1800 3850 1800 3863 1787 3863 storage-tank 1
-1769 3838 1783 3838 1783 3852 1769 3852 storage-tank 1
-2745 3022 2757 3022 2757 3034 2745 3034 storage-tank 1
-2276 3694 2289 3694 2289 3708 2276 3708 storage-tank 1
-2770 3031 2788 3031 2788 3050 2770 3050 storage-tank 0
-2903 2803 2918 2803 2918 2819 2903 2819 storage-tank 0
-2665 375 2650 399 2534 326 2547 305 ship 0
-2340 3703 2361 3703 2361 3722 2340 3722 storage-tank 0
-2314 3698 2333 3698 2333 3719 2314 3719 storage-tank 0
-2294 3698 2306 3698 2306 3711 2294 3711 storage-tank 1
-2280 3757 2305 3757 2305 3780 2280 3780 storage-tank 0
-1754 3836 1767 3836 1767 3849 1754 3849 storage-tank 1
-2247 3752 2271 3752 2271 3776 2247 3776 storage-tank 0
-2214 3745 2236 3745 2236 3771 2214 3771 storage-tank 0
-2181 3741 2204 3741 2204 3766 2181 3766 storage-tank 0
-1771 3927 1781 3927 1781 3937 1771 3937 storage-tank 1
-1769 3907 1786 3907 1786 3924 1769 3924 storage-tank 0
-1772 3893 1789 3893 1789 3909 1772 3909 storage-tank 0
-3145 3849 3163 3849 3163 3866 3145 3866 storage-tank 0
-3141 3871 3158 3871 3158 3886 3141 3886 storage-tank 0
-3276 3780 3291 3780 3291 3796 3276 3796 storage-tank 0
-2410 3701 2434 3701 2434 3727 2410 3727 storage-tank 0
-3356 513 3344 523 3299 488 3306 475 ship 0
-3438 396 3431 412 3385 387 3390 375 ship 0
-3328 428 3323 439 3263 403 3268 391 ship 0
-3033 2360 3028 2372 2967 2355 2971 2343 ship 0
-3018 2352 2994 2346 3019 2276 3036 2283 harbor 0
-2439 3707 2465 3707 2465 3733 2439 3733 storage-tank 0
-2093 49 2109 25 2261 118 2241 145 ship 0
-3847 558 3833 566 3795 510 3811 501 ship 0
-2531 278 2512 312 2320 193 2342 160 ship 0
-1732 1674 1719 1685 1666 1628 1681 1616 ship 0
-1746 1664 1735 1674 1680 1613 1695 1604 ship 0
-3887 3182 3951 3182 3951 3246 3887 3246 storage-tank 0
-3796 3166 3856 3166 3856 3223 3796 3223 storage-tank 0
-3903 3091 3966 3091 3966 3149 3903 3149 storage-tank 0
-3644 513 3635 522 3584 480 3593 471 ship 0
-3479 559 3473 568 3429 548 3435 537 ship 0
-3707 3143 3758 3143 3758 3196 3707 3196 storage-tank 0
-3830 3930 3890 3930 3890 3992 3830 3992 storage-tank 0
-1539 2050 1576 2041 1609 2064 1570 2077 bridge 0
-1792 2399 1832 2428 1804 2472 1762 2445 bridge 0
-38 2257 84 2240 79 2279 38 2291 bridge 0
-2004 2256 2012 2246 2139 2325 2130 2333 harbor 0
-2341 2443 2348 2431 2505 2513 2498 2523 harbor 0
-3166 2803 3169 2782 4000 2934 3999 2961 harbor 0
-3755 3913 3815 3913 3815 3976 3755 3976 storage-tank 0
-3549 592 3546 599 3500 574 3504 565 ship 0
-3185 2771 3266 2785 3262 2798 3182 2785 ship 0
-3427 2816 3421 2829 3351 2814 3353 2800 ship 0
-3625 2863 3628 2848 3724 2865 3722 2881 ship 0
-4001 771 3995 785 3926 732 3933 721 ship 0
-3829 835 3818 849 3740 801 3748 787 ship 0
-3534 596 3529 601 3503 587 3508 582 ship 0
-3813 3071 3875 3071 3875 3134 3813 3134 storage-tank 0
-3637 3132 3687 3132 3687 3186 3637 3186 storage-tank 0
-3297 3785 3311 3785 3311 3800 3297 3800 storage-tank 0
-3354 3861 3407 3861 3407 3912 3354 3912 storage-tank 0
-3648 3897 3699 3897 3699 3946 3648 3946 storage-tank 0
-3583 3884 3634 3884 3634 3933 3583 3933 storage-tank 0
-3519 3869 3568 3869 3568 3922 3519 3922 storage-tank 0
-3405 3939 3457 3939 3457 3990 3405 3990 storage-tank 0
-3339 3927 3392 3927 3392 3976 3339 3976 storage-tank 0
-3419 3874 3471 3874 3471 3926 3419 3926 storage-tank 0
-3245 3913 3295 3913 3295 3960 3245 3960 storage-tank 0
-3567 3948 3619 3948 3619 3997 3567 3997 storage-tank 0
-3173 3897 3224 3897 3224 3946 3173 3946 storage-tank 0
-3261 3839 3311 3839 3311 3891 3261 3891 storage-tank 0
-3191 3825 3238 3825 3238 3875 3191 3875 storage-tank 0
-3412 3769 3439 3769 3439 3799 3412 3799 storage-tank 0
-3383 3763 3409 3763 3409 3791 3383 3791 storage-tank 0
-3317 3790 3331 3790 3331 3805 3317 3805 storage-tank 0
-3502 3935 3555 3935 3555 3984 3502 3984 storage-tank 0
-3633 3961 3686 3961 3686 4000 3633 4000 storage-tank 0
-3716 3074 3774 3074 3774 3129 3716 3129 storage-tank 0
-3475 3018 3532 3018 3532 3075 3475 3075 storage-tank 0
-3651 3064 3704 3064 3704 3118 3651 3118 storage-tank 0
-3732 3007 3785 3007 3785 3060 3732 3060 storage-tank 0
-3669 3008 3710 3008 3710 3045 3669 3045 storage-tank 0
-3536 3110 3598 3110 3598 3174 3536 3174 storage-tank 0
-3457 3098 3518 3098 3518 3156 3457 3156 storage-tank 0
-3553 3035 3612 3035 3612 3092 3553 3092 storage-tank 0
-3568 2954 3629 2954 3629 3010 3568 3010 storage-tank 0
-3269 3062 3327 3062 3327 3121 3269 3121 storage-tank 0
-3487 2937 3549 2937 3549 2999 3487 2999 storage-tank 0
-3380 2919 3435 2919 3435 2974 3380 2974 storage-tank 0
-3298 2904 3358 2904 3358 2961 3298 2961 storage-tank 0
-3362 2996 3425 2996 3425 3056 3362 3056 storage-tank 0
-3284 2982 3342 2982 3342 3043 3284 3043 storage-tank 0
-3348 3074 3407 3074 3407 3135 3348 3135 storage-tank 0
-901 647 907 642 930 664 924 668 ship 0
