--- DOTA_devkit/example/labelTxt/P1888.txt
@@ -1,66 +1,0 @@
-imagesource:GoogleEarth
-gsd:0.266170468393
-674 375 683 375 684 394 675 395 small-vehicle 0
-683 351 694 351 695 371 684 371 small-vehicle 0
-465 371 455 372 451 324 460 323 large-vehicle 0
-451 374 441 375 436 327 445 326 large-vehicle 0
-439 373 429 374 422 327 433 326 large-vehicle 0
-426 376 415 378 408 330 418 328 large-vehicle 0
-411 377 400 379 395 332 405 330 large-vehicle 0
-398 377 388 379 382 331 392 330 large-vehicle 0
-384 382 374 383 368 334 377 333 large-vehicle 0
-370 382 360 383 355 336 364 335 large-vehicle 0
-354 384 345 383 341 337 352 336 large-vehicle 0
-328 387 318 388 312 342 323 340 large-vehicle 0
-316 389 307 392 300 344 308 342 large-vehicle 0
-302 390 292 392 286 343 296 342 large-vehicle 0
-288 394 278 394 273 346 281 345 large-vehicle 0
-531 213 520 215 518 180 528 179 large-vehicle 0
-517 213 507 215 502 174 512 172 large-vehicle 0
-504 216 494 217 490 176 498 174 large-vehicle 0
-490 216 481 217 474 176 484 174 large-vehicle 0
-480 245 470 246 466 205 475 203 large-vehicle 0
-467 248 457 249 453 207 462 205 large-vehicle 0
-454 248 445 249 441 208 450 207 large-vehicle 0
-441 250 431 251 427 209 437 208 large-vehicle 0
-419 253 409 254 402 212 410 210 large-vehicle 0
-382 259 372 261 366 219 376 218 large-vehicle 0
-368 260 359 261 353 218 362 217 large-vehicle 0
-356 259 347 261 339 220 348 219 large-vehicle 0
-343 262 334 265 327 222 335 221 large-vehicle 0
-319 266 310 268 304 225 314 223 large-vehicle 0
-307 268 297 269 292 227 301 226 large-vehicle 0
-294 261 284 264 277 221 287 220 large-vehicle 0
-480 369 469 371 463 321 473 320 large-vehicle 0
-493 367 483 368 476 321 486 319 large-vehicle 0
-507 365 498 368 489 320 499 318 large-vehicle 0
-457 442 468 443 470 485 461 486 large-vehicle 0
-681 369 672 369 671 351 680 351 small-vehicle 0
-654 374 664 374 665 395 654 396 small-vehicle 0
-697 281 706 280 707 300 698 300 small-vehicle 0
-687 281 697 280 698 299 688 299 small-vehicle 0
-677 282 687 282 688 302 678 303 small-vehicle 0
-667 286 676 285 677 303 669 304 small-vehicle 0
-657 287 666 286 667 306 657 307 small-vehicle 0
-696 263 705 261 707 280 698 280 small-vehicle 0
-686 260 695 259 695 278 687 278 small-vehicle 0
-665 262 675 261 676 280 667 281 small-vehicle 0
-656 262 665 263 666 284 656 284 small-vehicle 0
-645 267 654 267 655 285 646 285 small-vehicle 0
-469 442 479 441 485 484 474 485 large-vehicle 0
-443 444 454 444 458 487 448 488 large-vehicle 0
-520 362 511 364 503 316 513 315 large-vehicle 0
-430 446 440 445 446 489 436 490 large-vehicle 0
-419 450 430 450 436 495 425 497 large-vehicle 0
-406 451 416 451 421 495 410 495 large-vehicle 0
-394 454 403 453 409 496 400 497 large-vehicle 0
-381 455 393 455 396 496 386 498 large-vehicle 0
-368 452 379 452 383 496 373 497 large-vehicle 0
-354 455 365 454 371 497 360 498 large-vehicle 0
-343 455 354 455 358 498 347 498 large-vehicle 0
-316 458 327 456 335 499 325 502 large-vehicle 0
-304 463 315 461 320 505 310 506 large-vehicle 0
-293 463 303 462 308 505 298 506 large-vehicle 0
-281 466 291 464 296 508 286 508 large-vehicle 0
-200 449 210 450 210 493 200 492 large-vehicle 0
-197 200 207 200 208 242 198 243 large-vehicle 0
