--- DOTA_devkit/example/labelTxt/P2598.txt
@@ -1,28 +1,0 @@
-imagesource:GoogleEarth
-gsd:0.113068841441
-146 1655 152 1640 186 1654 180 1668 small-vehicle 0
-327 1328 344 1328 345 1369 328 1369 small-vehicle 0
-1395 526 2041 557 1997 1493 1354 1460 soccer-ball-field 0
-1365 255 2129 307 2028 1798 1289 1754 ground-track-field 0
-263 1526 364 1531 349 1746 248 1740 tennis-court 0
-536 1366 637 1372 626 1588 523 1580 tennis-court 0
-687 1375 790 1381 777 1597 675 1591 tennis-court 0
-477 1678 689 1692 683 1792 468 1779 tennis-court 0
-1080 110 1079 125 1044 121 1045 105 small-vehicle 0
-1124 108 1122 122 1091 122 1090 107 small-vehicle 0
-1175 106 1173 125 1140 123 1139 105 small-vehicle 0
-1381 106 1380 125 1342 122 1341 105 small-vehicle 0
-1438 104 1438 123 1396 122 1396 102 small-vehicle 0
-1445 123 1444 106 1481 106 1482 123 small-vehicle 0
-1597 106 1597 123 1557 123 1557 104 small-vehicle 0
-1550 103 1551 124 1509 126 1508 107 small-vehicle 0
-1868 65 1865 44 1930 43 1933 65 large-vehicle 0
-1881 100 1879 117 1843 117 1841 98 small-vehicle 0
-1945 97 1946 116 1898 118 1897 100 small-vehicle 0
-1429 1 1446 2 1444 33 1430 33 small-vehicle 0
-1520 0 1538 0 1540 41 1521 42 small-vehicle 0
-1452 1 1470 1 1470 40 1452 39 small-vehicle 0
-2035 58 2034 42 2074 42 2076 58 small-vehicle 0
-168 1231 167 1248 115 1247 116 1226 large-vehicle 0
-289 1315 315 1316 315 1373 290 1374 large-vehicle 0
-606 337 1265 367 1245 843 582 814 soccer-ball-field 0
